# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Story Generation Workflow API
  version: 1.0.0
  description: |
    Design-first specification for the Story Generation Workflow service. This document captures
    the currently implemented Express route handlers and internal services that orchestrate story
    creation, editing, media generation, audiobook production, and workflow coordination.
servers:
  - url: https://story-generation-workflow.example.com
    description: Production
  - url: http://localhost:8080
    description: Local development
security:
  - {}
...
paths:
...
  /api/jobs/text-edit:
    post:
      tags: [Jobs]
      summary: Create a text edit job
      description: Creates an asynchronous job that edits a single chapter or entire story using background workers.
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TextEditJobRequest'
      responses:
        '200':
          description: Job created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobCreationResponse'
        '400':
          description: Validation error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Job creation failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/jobs/image-edit:
    post:
      tags: [Jobs]
      summary: Create an image edit job
      description: Enqueues an asynchronous job that edits or converts a story illustration.
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImageEditJobRequest'
      responses:
        '200':
          description: Job created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobCreationResponse'
        '400':
          description: Validation error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Job creation failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/jobs/translate-text:
    post:
      tags: [Jobs]
      summary: Create a translation job
      description: Creates a background job that translates an entire story into the specified locale.
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TranslationJobRequest'
      responses:
        '200':
          description: Job created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobCreationResponse'
        '400':
          description: Validation error or locale conflict.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Story not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Job creation failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/jobs/{jobId}:
    get:
      tags: [Jobs]
      summary: Get job status
      description: Returns current status, simulated progress, metadata, and optional result for an asynchronous job.
      security:
        - ApiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/JobIdParam'
      responses:
        '200':
          description: Job status returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobStatusResponse'
        '400':
          description: Missing job identifier.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Job not found or expired.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Status lookup failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
  parameters:
...
    JobIdParam:
      name: jobId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Unique identifier returned when the job was created.
...
  schemas:
...
    ErrorResponse:
      type: object
      description: Standard error envelope returned across the API. Fields may vary slightly by route.
      properties:
        success:
          type: boolean
          description: Indicates success when present. Typically omitted for internal-only endpoints that predate the convention.
          example: false
        error:
          type: string
          description: Human-readable error description.
        message:
          type: string
          description: Optional detailed error message.
        details:
          type: object
          additionalProperties: true
          description: Optional machine-readable context.
      required:
        - error
...
    TextEditJobRequest:
      type: object
      required:
        - storyId
        - userRequest
      properties:
        storyId:
          type: string
          format: uuid
        userRequest:
          type: string
          maxLength: 2000
        scope:
          type: string
          enum: [chapter, story]
          default: chapter
        chapterNumber:
          type: integer
          minimum: 1
    ImageEditJobRequest:
      type: object
      required:
        - storyId
        - imageUrl
        - imageType
      properties:
        storyId:
          type: string
          format: uuid
        imageUrl:
          type: string
        imageType:
          type: string
          enum: [cover, backcover, chapter]
        userRequest:
          type: string
        chapterNumber:
          type: integer
          minimum: 1
        graphicalStyle:
          type: string
        userImageUri:
          type: string
        convertToStyle:
          type: boolean
    TranslationJobRequest:
      type: object
      required:
        - storyId
        - targetLocale
      properties:
        storyId:
          type: string
          format: uuid
        targetLocale:
          type: string
          enum:
            - en-US
            - en-GB
            - pt-PT
            - pt-BR
            - es-ES
            - fr-FR
            - it-IT
            - de-DE
            - nl-NL
            - pl-PL
    JobCreationResponse:
      type: object
      properties:
        success:
          type: boolean
        jobId:
          type: string
          format: uuid
        estimatedDuration:
          type: integer
          description: Estimated duration in milliseconds.
        message:
          type: string
    JobStatus:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
        status:
          type: string
          enum: [pending, processing, completed, failed]
        progress:
          type: number
        elapsedTime:
          type: integer
        remainingTime:
          type: integer
        estimatedDuration:
          type: integer
        metadata:
          type: object
          additionalProperties: true
        result:
          type: object
          additionalProperties: true
        error:
          type: string
    JobStatusResponse:
      type: object
      properties:
        success:
          type: boolean
        job:
          $ref: '#/components/schemas/JobStatus'
...